requirements.txt
fastapi~=0.115.12
uvicorn
sqlalchemy~=2.0.40
psycopg2
qrcode~=8.1
reportlab~=4.3.1
python-dotenv~=1.1.0
requests  # For Paystack API calls
pydantic  # For data validation

pillow~=11.1.0